+++ src/main/java/de/maxhenkel/miningdimension/IItemBlock.java
@@ -1,9 +1,0 @@
-package de.maxhenkel.miningdimension;
-
-import net.minecraft.item.Item;
-
-public interface IItemBlock {
-
-    Item toItem();
-
-}