+++ src/main/java/de/maxhenkel/miningdimension/dimension/ChunkGeneratorMining.java
@@ -1,22 +1,46 @@
 package de.maxhenkel.miningdimension.dimension;
 
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.Blockreader;
+import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorld;
 import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.chunk.IChunk;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.Heightmap;
-import net.minecraft.world.gen.OverworldGenSettings;
 import net.minecraft.world.gen.WorldGenRegion;
+import net.minecraft.world.gen.feature.structure.StructureManager;
+import net.minecraft.world.gen.settings.DimensionStructuresSettings;
+
+import java.util.Collections;
+import java.util.Optional;
 
 public class ChunkGeneratorMining extends ChunkGenerator {
 
+    public static final Codec<ChunkGeneratorMining> CODEC = RecordCodecBuilder.create((instance) ->
+            instance.group(BiomeProvider.field_235202_a_.fieldOf("biome_source").forGetter((generator) -> generator.biomeProvider)
+            ).apply(instance, instance.stable(ChunkGeneratorMining::new))
+    );
+
     private int height;
 
-    public ChunkGeneratorMining(IWorld world, BiomeProvider biomeProvider, OverworldGenSettings settings) {
-        super(world, biomeProvider, settings);
+    public ChunkGeneratorMining(BiomeProvider biomeProvider) {
+        super(biomeProvider, new DimensionStructuresSettings(Optional.empty(), Collections.emptyMap()));
         height = 255;
+    }
+
+    @Override
+    protected Codec<? extends ChunkGenerator> func_230347_a_() {
+        return CODEC;
+    }
+
+    @Override
+    public ChunkGenerator func_230349_a_(long l) {
+        return this;
     }
 
     @Override
@@ -30,12 +54,7 @@ public class ChunkGeneratorMining extends ChunkGenerator {
     }
 
     @Override
-    public int getGroundHeight() {
-        return height;
-    }
-
-    @Override
-    public void makeBase(IWorld world, IChunk chunk) {
+    public void func_230352_b_(IWorld world, StructureManager structureManager, IChunk chunk) {
         BlockPos.Mutable pos = new BlockPos.Mutable();
 
         for (int x = 0; x < 16; x++) {
@@ -54,7 +73,18 @@ public class ChunkGeneratorMining extends ChunkGenerator {
     }
 
     @Override
-    public int func_222529_a(int i, int i1, Heightmap.Type type) {
+    public int func_222529_a(int i1, int i2, Heightmap.Type heightmapType) {
         return 0;
     }
+
+    @Override
+    public IBlockReader func_230348_a_(int i1, int i2) {
+        return new Blockreader(new BlockState[]{Blocks.STONE.getDefaultState()});
+    }
+
+    @Override
+    public int getGroundHeight() {
+        return height;
+    }
+
 }