+++ changelog.md
@@ -1,1 +1,2 @@
-- Fixed teleporter not working in worlds created before adding the mod
+- Added german language
+- Updated the teleporter texture to match the new textures