+++ changelog.md
@@ -1,2 +1,1 @@
-- Added german language
-- Updated the teleporter texture to match the new textures
+- Update to 1.16.1